requirement.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cloudinary==1.15.0
Flask==1.0.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
mock==2.0.0
numpy==1.16.2
opencv-python==4.0.0.21
pbr==5.1.3
Pillow==5.4.1
pytesseract==0.2.6
requests==2.21.0
six==1.12.0
tika==1.19
urllib3==1.24.1
Werkzeug==0.15.1
wino==0.0.2
